+++ crates/ruma-state-res/CHANGELOG.md
@@ -3,6 +3,7 @@
 Breaking changes:
 
 * state_res::resolve auth_events type has been slightly changed and renamed to auth_chain_sets
+* state_res::resolve structs were changed from BTreeMap/Set to HashMap/Set
 
 # 0.2.0